replaceParams should escape \ characters
When using replaceParams to insert a string into a SQL string, it should
escape backslashes as a pair of backslashes.

replaceParams escapes " characters as \".

Before this change, if the input string was
\"
then the output string would be
\\"
which would mean the " character was no longer escaped.

https://dev.mysql.com/doc/refman/8.0/en/string-literals.html#character-escape-sequences

This commit includes a few small other changes to replaceParams.
Where `indexOf` was used to determine if a string contained a pattern
I've replaced it with `includes`. This is better for people like me who
aren't primarily Javascript/Typescript programmers.

-        (val.toLowerCase().indexOf(" as ") > -1 ||
-          val.toLowerCase().indexOf(" AS ") > -1)

The 2nd condition was never true because if the string is forced lower
case it will never contain some uppper case characters.

https://github.com/manyuanrong/deno_mysql/issues/57